ci: Build crates individually
For some crates, like the GTK PaintableSink or the D3D sink,
we can't build them with --all-features as some of them are
conflicting, like enabling x11/wayland on windows builds
and vice versa.

Thus we need to build the crates individually to have more
over the features enabled in each build.