spritesheet: Element to generate spritesheet.
spritesheet are used for thumbnail seeking in the player.
It is used along with webvtt file, defined as sparial-dimension
of media. https://www.w3.org/TR/media-frags/#naming-space

This plugin only generate the output buffer with rows and columns
of tiles(thumbnails).
It sends a signal on every tile on spritesheert. Signal has all
the infomation to generate entry in the webvtt file.