livesync: add support for variable framerate
A variable framerate is signaled by a framerate of 0/1 and the
max-framerate field. This may be the result of source that can produce
up to max-framerate frames, but may produce less. A live stream may
expect frames with the expected max-framerate and updated timestamps
even if the frame contents didn't change.

Support streams with variable framerate in livesync and send frames with
max-framerate to downstream.