fmp4mux: allow targeting a minimum fragment duration
Use case:

* We know maximum keyframe interval from input stream (eg 2 seconds)
* We want to target a minimum fragment duration (eg 5 seconds)

We can thus use latency 2 seconds, and ensure that we always have
enough time to prepare 6 seconds fragments, but we also don't need
to set fragment-duration to 7 seconds to achieve that.

That means that if we actually get one keyframe per second from the
input stream, we can output the optimal 5-second duration fragments.