video/gtk4: Add dmabuf to the default features
We only ever use this feature along with checking for
target platform=linux, which means it doesn't have any
effect on macos and windows and thus we could try
to be building the feature by default if we are on linux.

Previously the dmabuf feature depended on the wayland feature
but that would try to build wayland crates on windows/macos and
thus we need to remove that. In practice wayland is only a runtime
dependency as the only working dmabuf platform, and the dmabuf
negotiation will happen automatically if the proper conditions
are met. (Running on wayland and with a Wayland context and having
capable hardware and so on).