meson: Enable statically linking Rust plugins under MSVC
This performs an abomination of a surgery in the cargo-c static
libraries... namely, it extracts all the common members
(libcompiler_builtins and libstd) out of the final artifacts, exposes
them as compiler flags, and then relinks them when producing the final
artifacts. Special note must be made of certain import descriptor that
Rust internally uses to map synchronization and RNG primitives -- these
aren't .o files, so they must be carried over verbatim.

This requires the Rust llvm-tools component, as the only uncovered
function is `rust_eh_personality`, and only objcopy is able to rename
symbols for PE objects.

This also has the side effect that .pc files are now hardcoded to the
Rust version used to compile the binary.

Fixes #600