rtp: tests: add Source::from_gdp_data()
That grabs caps and buffers from a GDP stream, for easier
unit testing data generation (especially useful for formats
that are not parsable stand-alone or have wacky caps or
timestamps).

In future we could add a custom source element that can
inject all the events as well, which would be useful for
simulating iput caps changes, but for now we just implement
what we need.