hlsbasesink: Avoid unnecesary media sequence bump when closing playlist
update_playlist_state() assumes it's called when a fragment is added,
but we also call it when closing the playlist. That results in playlists
which don't have the media sequence number specified (so it's 0) and
then suddenly it's written as 1 when the playlist is closed.

With this change, it's 1 from the beginning and won't be changed unless
fragments are removed of course.