hlssink3: Fix EXT-X-MEDIA-SEQUENCE calculation
This was due to where the internal playlist index value was updated, it should
only be incremented when adding actual segments to the playlist.

The media playlist media_sequence value is then always the last
playlist_index (i.e. the sequence number for the last added entry) minus the
number of entries in the playlist. Regardless of the number of times you update
the media playlist.