awss3sink: added matroska seeking test
As identified, this element seeks many times which causes the
cache to break for some reason (still broken in this commit, and
this test may or may not be 100% correct).  In this commit, the
writer harness fails during some of the seekin because it asserts
that the seek went out-of-bounds.