awss3sink: addressed seek-to-end test behavior
As identified, the mostraska element's seeking behavior would
cause these changes to fail and exposed some extra corner cases
involving part boundary edges that are addressed with a new test
and several changes in this patch.  The test writes 3 parts with
head cache of 1.  It verifies that after having written into part
2 that part 1 is still accessible, then writes through part 2,
thus beginning part 3 (but with no bytes).  It then verifies that
in having done so, it can seek into 1 and write, as well as seek
to just after the end of part 2 (part 3 start byte 0) and write
but cannot seek into the non-cached part 2 (the seek works as I
think we have no choice but to allow given the gap, but it will
cache miss on the write with an error, which is the right overall
behavior).  The test concludes by downloading the result and doing
a byte-by-byte comparison of the expected file output and the read.