fixup: rtp: vorbis: config: simplify VorbisHeaders::from_streamheaders() and accept big vorbis comments
VorbisHeaders is just the various vorbis headers in a Vec, plus
a VorbisInfo struct with the metadata needed for caps. There's no
limitation on the size of the vorbis headers in this representation.

Avoid the unnecessary roundtrip via ::pack_headers() which does check
for a max 16-bit total size, which is not warranted here.